requirements.txt
django>4.1, <5.2
psycopg2
Pillow
sorl-thumbnail